refactor: Reduce cognitive complexity
