+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - chore(deps): Update package-lock files to NPM v8
 - test(e2e): Increase tests stability. Fix flaky test in parallel specs
+- test(e2e): Turn tsc check into an assertion
+- refactor: Reduce cognitive complexity
 
 ## [3.3.0] - 2021-11-13
