+++ src/helpers/constants.js
@@ -13,41 +13,11 @@ const STOP_MESSAGE = "Stopping Cypress runner due to a previous failure";
 const SKIP_MESSAGE = "Enabling skip mode";
 const FAILED_TEST_MESSAGE = "Enabling skip mode";
 
-const STRATEGIES = ["spec", "parallel"];
-
-const TRUTHY_VALUES = [true, "true", 1, "1"];
-const FALSY_VALUES = [false, "false", 0, "0"];
-const UNDEFINED_VALUES = ["", undefined];
-
 const ENVIRONMENT_DEFAULT_VALUES = {
   [PLUGIN_ENVIRONMENT_VAR]: true,
   [ENABLED_ENVIRONMENT_VAR]: true,
   [BAIL_ENVIRONMENT_VAR]: 1,
 };
-
-function valueIsOneOf(value, arrayOfValues) {
-  return arrayOfValues.includes(value);
-}
-
-function strategyIsParallel(value) {
-  return value === STRATEGIES[1];
-}
-
-function strategyIsSpec(value) {
-  return value === STRATEGIES[0];
-}
-
-function isTruthy(value) {
-  return valueIsOneOf(value, TRUTHY_VALUES);
-}
-
-function isFalsy(value) {
-  return valueIsOneOf(value, FALSY_VALUES);
-}
-
-function isDefined(value) {
-  return !valueIsOneOf(value, UNDEFINED_VALUES);
-}
 
 module.exports = {
   ENVIRONMENT_DEFAULT_VALUES,
@@ -63,9 +33,4 @@ module.exports = {
   STOP_MESSAGE,
   SKIP_MESSAGE,
   FAILED_TEST_MESSAGE,
-  isTruthy,
-  isFalsy,
-  isDefined,
-  strategyIsParallel,
-  strategyIsSpec,
 };
