+++ src/support.js
@@ -1,9 +1,4 @@
 const {
-  ENVIRONMENT_DEFAULT_VALUES,
-  PLUGIN_ENVIRONMENT_VAR,
-  ENABLED_ENVIRONMENT_VAR,
-  STRATEGY_ENVIRONMENT_VAR,
-  BAIL_ENVIRONMENT_VAR,
   SHOULD_SKIP_TASK,
   RESET_SKIP_TASK,
   FAILED_TESTS_TASK,
@@ -12,112 +7,21 @@ const {
   STOP_MESSAGE,
   SKIP_MESSAGE,
   FAILED_TEST_MESSAGE,
-  isFalsy,
-  isTruthy,
-  isDefined,
-  strategyIsSpec,
-} = require("./helpers");
+} = require("./helpers/constants");
+
+const {
+  bailConfig,
+  pluginIsEnabled,
+  failFastIsEnabled,
+  currentStrategyIsSpec,
+} = require("./helpers/config");
+
+const { isHeaded, testHasFailed, wrapCypressRunner, stopRunner } = require("./helpers/cypress");
 
 function support(Cypress, cy, beforeEach, afterEach, before) {
-  let hookFailed, hookFailedName, hookError;
-
-  function isHeaded() {
-    return Cypress.browser && Cypress.browser.isHeaded;
-  }
-
-  function booleanEnvironmentVarValue(environmentVarName) {
-    const defaultValue = ENVIRONMENT_DEFAULT_VALUES[environmentVarName];
-    const value = Cypress.env(environmentVarName);
-    const isTruthyValue = isTruthy(value);
-    if (!isTruthyValue && !isFalsy(value)) {
-      return defaultValue;
-    }
-    return isTruthyValue;
-  }
-
-  function numberEnvironmentVarValue(environmentVarName) {
-    const defaultValue = ENVIRONMENT_DEFAULT_VALUES[environmentVarName];
-    const value = Cypress.env(environmentVarName);
-    if (isDefined(value)) {
-      return Number(value);
-    }
-    return defaultValue;
-  }
-
-  function getFailFastEnvironmentConfig() {
-    return {
-      plugin: booleanEnvironmentVarValue(PLUGIN_ENVIRONMENT_VAR),
-      enabled: booleanEnvironmentVarValue(ENABLED_ENVIRONMENT_VAR),
-      strategyIsSpec: strategyIsSpec(Cypress.env(STRATEGY_ENVIRONMENT_VAR)),
-      bail: numberEnvironmentVarValue(BAIL_ENVIRONMENT_VAR),
-    };
-  }
-
-  function getTestConfig(test) {
-    // Cypress <6.7
-    if (test.cfg) {
-      return test.cfg;
-    }
-    // Cypress >9
-    if (
-      test.ctx &&
-      test.ctx.test &&
-      test.ctx.test._testConfig &&
-      test.ctx.test._testConfig.testConfigList &&
-      test.ctx.test._testConfig.testConfigList[
-        test.ctx.test._testConfig.testConfigList.length - 1
-      ] &&
-      test.ctx.test._testConfig.testConfigList[test.ctx.test._testConfig.testConfigList.length - 1]
-        .overrides
-    ) {
-      return test.ctx.test._testConfig.testConfigList[
-        test.ctx.test._testConfig.testConfigList.length - 1
-      ].overrides;
-    }
-    // Cypress >6.7
-    if (test.ctx && test.ctx.test && test.ctx.test._testConfig) {
-      return test.ctx.test._testConfig;
-    }
-    return {};
-  }
-
-  function getTestFailFastConfig(currentTest) {
-    const testConfig = getTestConfig(currentTest);
-    if (testConfig.failFast) {
-      return {
-        ...getFailFastEnvironmentConfig(),
-        ...testConfig.failFast,
-      };
-    }
-    if (currentTest.parent) {
-      return getTestFailFastConfig(currentTest.parent);
-    }
-    return getFailFastEnvironmentConfig();
-  }
-
-  function currentStrategyIsSpec() {
-    return getFailFastEnvironmentConfig().strategyIsSpec;
-  }
-
-  function pluginIsEnabled() {
-    return getFailFastEnvironmentConfig().plugin;
-  }
-
-  function bailConfig() {
-    return getFailFastEnvironmentConfig().bail;
-  }
-
-  function failFastIsEnabled(currentTest) {
-    return getTestFailFastConfig(currentTest).enabled;
-  }
-
-  function testHasFailed(currentTest) {
-    return currentTest.state === "failed" && currentTest.currentRetry() === currentTest.retries();
-  }
-
   function stopCypressRunner() {
     cy.task(LOG_TASK, STOP_MESSAGE);
-    Cypress.runner.stop();
+    stopRunner(Cypress);
   }
 
   function resetSkipFlag() {
@@ -136,7 +40,7 @@ function support(Cypress, cy, beforeEach, afterEach, before) {
   function registerFailureAndRunIfBailLimitIsReached(callback) {
     cy.task(LOG_TASK, FAILED_TEST_MESSAGE);
     cy.task(FAILED_TESTS_TASK, true, { log: false }).then((value) => {
-      if (value >= bailConfig()) {
+      if (value >= bailConfig(Cypress)) {
         callback();
       }
     });
@@ -151,7 +55,7 @@ function support(Cypress, cy, beforeEach, afterEach, before) {
   }
 
   beforeEach(function () {
-    if (pluginIsEnabled()) {
+    if (pluginIsEnabled(Cypress)) {
       runIfSkipIsEnabled(() => {
         this.currentTest.pending = true;
         stopCypressRunner();
@@ -164,9 +68,9 @@ function support(Cypress, cy, beforeEach, afterEach, before) {
     const currentTest = this.currentTest;
     if (
       currentTest &&
-      pluginIsEnabled() &&
+      pluginIsEnabled(Cypress) &&
       testHasFailed(currentTest) &&
-      failFastIsEnabled(currentTest)
+      failFastIsEnabled(currentTest, Cypress)
     ) {
       registerFailureAndRunIfBailLimitIsReached(() => {
         enableSkipMode();
@@ -175,13 +79,10 @@ function support(Cypress, cy, beforeEach, afterEach, before) {
   });
 
   before(function () {
-    if (pluginIsEnabled()) {
-      if (isHeaded() || currentStrategyIsSpec()) {
+    if (pluginIsEnabled(Cypress)) {
+      if (isHeaded(Cypress) || currentStrategyIsSpec(Cypress)) {
         /*
-          Reset the shouldSkip flag at the start of a run, so that it
-          doesn't carry over into subsequent runs.
-          Do this only for headed runs because in headless runs,
-          the `before` hook is executed for each spec file.
+          Reset the shouldSkip flag at the start of a run, so that it doesn't carry over into subsequent runs. Do this only for headed runs because in headless runs, the `before` hook is executed for each spec file.
         */
         resetSkipFlag();
         resetFailedTests();
@@ -193,46 +94,8 @@ function support(Cypress, cy, beforeEach, afterEach, before) {
     }
   });
 
-  const _onRunnableRun = Cypress.runner.onRunnableRun;
-
-  if (pluginIsEnabled()) {
-    Cypress.runner.onRunnableRun = function (runnableRun, runnable, args) {
-      const isHook = runnable.type === "hook";
-      const isBeforeHook = isHook && runnable.hookName.match(/before/);
-
-      const next = args[0];
-
-      const setFailedFlag = function (error) {
-        if (error) {
-          hookFailedName = runnable.hookName;
-          hookError = error;
-          hookFailed = true;
-        }
-        /* 
-          Do not pass the error, because Cypress stops if there is an error on before hooks,
-          so this plugin can't set the skip flag
-        */
-        return next.call(this /*, error */);
-      };
-
-      const forceTestToFail = function () {
-        hookFailed = false;
-        hookError.message = `"${hookFailedName}" hook failed: ${hookError.message}`;
-        // Force next test to fail, so the plugin can set the skip flag, and the test is marked as failed
-        return next.call(this, hookError);
-      };
-
-      if (isBeforeHook && hookFailed) {
-        // Skip other before hooks when one failed
-        return next();
-      } else if (!isHook && hookFailed) {
-        args[0] = forceTestToFail;
-      } else if (isBeforeHook) {
-        args[0] = setFailedFlag;
-      }
-
-      return _onRunnableRun.apply(this, [runnableRun, runnable, args]);
-    };
+  if (pluginIsEnabled(Cypress)) {
+    wrapCypressRunner(Cypress);
   }
 }
